Reduce dupplication with default interface implementation
